backend/requirements.txt
flask
flask-cors
python-dotenv
mongoengine
pymongo
pandas
joblib
PyJWT
werkzeug
flask-bcrypt